Rolled back test figure move to own file for linting issues
